+++ lambdas/functions/control-plane/src/github/auth.ts
@@ -10,6 +10,17 @@ import {
 import { OctokitOptions } from '@octokit/core/dist-types/types';
 import { request } from '@octokit/request';
 import { Octokit } from '@octokit/rest';
+<<<<<<< HEAD:lambdas/functions/control-plane/src/github/auth.ts
+import { throttling } from '@octokit/plugin-throttling';
+import { createChildLogger } from '@aws-github-runner/aws-powertools-util';
+import { getParameter } from '@aws-github-runner/aws-ssm-util';
+import { EndpointDefaults } from '@octokit/types';
+
+const logger = createChildLogger('gh-auth');
+
+export async function createOctokitClient(token: string, ghesApiUrl = ''): Promise<Octokit> {
+  const CustomOctokit = Octokit.plugin(throttling);
+=======
 import { createChildLogger } from '@terraform-aws-github-runner/aws-powertools-util';
 import { getParameter } from '@terraform-aws-github-runner/aws-ssm-util';
 
@@ -17,15 +28,32 @@ import { axiosFetch } from '../axios/fetch-override';
 
 const logger = createChildLogger('gh-auth');
 export async function createOctoClient(token: string, ghesApiUrl = ''): Promise<Octokit> {
+>>>>>>> main:lambdas/functions/control-plane/src/gh-auth/gh-auth.ts
   const ocktokitOptions: OctokitOptions = {
     auth: token,
-    request: { fetch: axiosFetch },
   };
   if (ghesApiUrl) {
     ocktokitOptions.baseUrl = ghesApiUrl;
     ocktokitOptions.previews = ['antiope'];
   }
+<<<<<<< HEAD:lambdas/functions/control-plane/src/github/auth.ts
+
+  return new CustomOctokit({
+    ...ocktokitOptions,
+    throttle: {
+      onRateLimit: (retryAfter: number, options: Required<EndpointDefaults>) => {
+        logger.warn(
+          `GitHub rate limit: Request quota exhausted for request ${options.method} ${options.url}. Requested `,
+        );
+      },
+      onSecondaryRateLimit: (retryAfter: number, options: Required<EndpointDefaults>) => {
+        logger.warn(`GitHub rate limit: SecondaryRateLimit detected for request ${options.method} ${options.url}`);
+      },
+    },
+  });
+=======
   return new Octokit(ocktokitOptions);
+>>>>>>> main:lambdas/functions/control-plane/src/gh-auth/gh-auth.ts
 }
 
 export async function createGithubAppAuth(
@@ -66,12 +94,7 @@ async function createAuth(installationId: number | undefined, ghesApiUrl: string
   if (ghesApiUrl) {
     authOptions.request = request.defaults({
       baseUrl: ghesApiUrl,
-      request: {
-        fetch: axiosFetch,
-      },
     });
-  } else {
-    authOptions.request = request.defaults({ request: { fetch: axiosFetch } });
   }
   return createAppAuth(authOptions);
 }