+++ lambdas/functions/control-plane/src/pool/pool.ts
@@ -1,10 +1,10 @@
 import { Octokit } from '@octokit/rest';
-import { createChildLogger } from '@terraform-aws-github-runner/aws-powertools-util';
+import { createChildLogger } from '@aws-github-runner/aws-powertools-util';
 import yn from 'yn';
 
 import { bootTimeExceeded, listEC2Runners } from '../aws/runners';
 import { RunnerList } from '../aws/runners.d';
-import { createGithubAppAuth, createGithubInstallationAuth, createOctoClient } from '../gh-auth/gh-auth';
+import { createGithubAppAuth, createGithubInstallationAuth, createOctokitClient } from '../github/auth';
 import { createRunners } from '../scale-runners/scale-up';
 
 const logger = createChildLogger('pool');
@@ -50,7 +50,7 @@ export async function adjust(event: PoolEvent): Promise<void> {
 
   const installationId = await getInstallationId(ghesApiUrl, runnerOwner);
   const ghAuth = await createGithubInstallationAuth(installationId, ghesApiUrl);
-  const githubInstallationClient = await createOctoClient(ghAuth.token, ghesApiUrl);
+  const githubInstallationClient = await createOctokitClient(ghAuth.token, ghesApiUrl);
 
   // Get statusses of runners registed in GitHub
   const runnerStatusses = await getGitHubRegisteredRunnnerStatusses(
@@ -110,7 +110,7 @@ export async function adjust(event: PoolEvent): Promise<void> {
 
 async function getInstallationId(ghesApiUrl: string, org: string): Promise<number> {
   const ghAuth = await createGithubAppAuth(undefined, ghesApiUrl);
-  const githubClient = await createOctoClient(ghAuth.token, ghesApiUrl);
+  const githubClient = await createOctokitClient(ghAuth.token, ghesApiUrl);
 
   return (
     await githubClient.apps.getOrgInstallation({