+++ lambdas/functions/termination-watcher/src/termination-warning.test.ts
@@ -3,11 +3,16 @@ import { mockClient } from 'aws-sdk-client-mock';
 import 'aws-sdk-client-mock-jest';
 import { handle } from './termination-warning';
 import { SpotInterruptionWarning, SpotTerminationDetail } from './types';
+<<<<<<< HEAD
+import { createSingleMetric } from '@aws-github-runner/aws-powertools-util';
+import { MetricUnit } from '@aws-lambda-powertools/metrics';
+=======
 import { createSingleMetric } from '@terraform-aws-github-runner/aws-powertools-util';
 import { MetricUnits } from '@aws-lambda-powertools/metrics';
+>>>>>>> main
 
-jest.mock('@terraform-aws-github-runner/aws-powertools-util', () => ({
-  ...jest.requireActual('@terraform-aws-github-runner/aws-powertools-util'),
+jest.mock('@aws-github-runner/aws-powertools-util', () => ({
+  ...jest.requireActual('@aws-github-runner/aws-powertools-util'),
   // eslint-disable-next-line @typescript-eslint/no-unused-vars
   createSingleMetric: jest.fn((name: string, unit: string, value: number, dimensions?: Record<string, string>) => {
     return {