+++ lambdas/functions/webhook/src/lambda.ts
@@ -1,15 +1,15 @@
 import middy from '@middy/core';
-import { logger, setContext, captureLambdaHandler, tracer } from '@terraform-aws-github-runner/aws-powertools-util';
+import { logger, setContext, captureLambdaHandler, tracer } from '@aws-github-runner/aws-powertools-util';
 import { APIGatewayEvent, Context } from 'aws-lambda';
 
 import { handle } from './webhook';
 import { Config } from './ConfigResolver';
 import { IncomingHttpHeaders } from 'http';
-import ValidationError from './ValidatonError';
+import ValidationError from './ValidationError';
 
 export interface Response {
   statusCode: number;
-  body?: string;
+  body: string;
 }
 
 middy(githubWebhook).use(captureLambdaHandler(tracer));