+++ modules/multi-runner/runners.tf
@@ -31,7 +31,7 @@ module "runners" {
   ami_id_ssm_parameter_name = each.value.runner_config.ami_id_ssm_parameter_name
   ami_kms_key_arn           = each.value.runner_config.ami_kms_key_arn
 
-  sqs_build_queue                      = { "arn" : each.value.arn }
+  sqs_build_queue                      = { "arn" : each.value.arn, "url" : each.value.url }
   github_app_parameters                = local.github_app_parameters
   ebs_optimized                        = each.value.runner_config.ebs_optimized
   enable_on_demand_failover_for_errors = each.value.runner_config.enable_on_demand_failover_for_errors
@@ -110,4 +110,8 @@ module "runners" {
   associate_public_ipv4_address              = var.associate_public_ipv4_address
 
   ssm_housekeeper = var.runners_ssm_housekeeper
+
+  job_retry = each.value.runner_config.job_retry
+
+  metrics = var.metrics
 }