+++ modules/multi-runner/termination-watcher.tf
@@ -13,9 +13,13 @@ locals {
     logging_retention_in_days = var.logging_retention_in_days
     role_path                 = var.role_path
     role_permissions_boundary = var.role_permissions_boundary
-    metrics_namespace         = var.metrics_namespace
     s3_bucket                 = var.lambda_s3_bucket
     tracing_config            = var.tracing_config
+<<<<<<< HEAD
+    lambda_tags               = var.lambda_tags
+    metrics                   = var.metrics
+=======
+>>>>>>> main
   }
 }